--- .changeset/old-paws-smoke.md
@@ -1,5 +1,0 @@
----
-"siws-better-auth": minor
----
-
-First release of the better-auth plugin
