+++ .changeset/slow-paws-show.md
@@ -1,0 +1,5 @@
+---
+"sign-in-with-stacks": minor
+---
+
+Add better-auth plugin
