+++ packages/sign-in-with-stacks/package.json
@@ -23,6 +23,18 @@
       "types": "./dist/index.d.ts",
       "default": "./dist/index.js"
     },
+    "./plugins": {
+      "types": "./dist/plugins/index.d.ts",
+      "default": "./dist/plugins/index.js"
+    },
+    "./plugins/better-auth": {
+      "types": "./dist/plugins/better-auth/index.d.ts",
+      "default": "./dist/plugins/better-auth/index.js"
+    },
+    "./plugins/better-auth/client": {
+      "types": "./dist/plugins/better-auth/client.d.ts",
+      "default": "./dist/plugins/better-auth/client.js"
+    },
     "./package.json": "./package.json"
   },
   "files": [
@@ -43,15 +55,27 @@
     "@stacks/transactions": "7.3.0",
     "@tsconfig/recommended": "1.0.13",
     "@types/node": "24.10.1",
+    "better-auth": "1.4.6",
     "publint": "0.3.16",
     "tsdown": "0.17.3",
     "typescript": "5.9.3",
-    "vitest": "4.0.15"
+    "vitest": "4.0.15",
+    "zod": "4.1.13"
   },
   "peerDependencies": {
     "@stacks/common": "^7.0.2",
     "@stacks/encryption": "^7.0.2",
     "@stacks/network": "^7.0.2",
-    "@stacks/transactions": "^7.0.2"
+    "@stacks/transactions": "^7.0.2",
+    "better-auth": "^1.3.0",
+    "zod": "^3.0.0 || ^4.0.0"
+  },
+  "peerDependenciesMeta": {
+    "better-auth": {
+      "optional": true
+    },
+    "zod": {
+      "optional": true
+    }
   }
 }
