+++ packages/sign-in-with-stacks/src/plugins/better-auth/index.ts
@@ -1,1 +1,2 @@
+export * from "./client.js";
 export * from "./plugin.js";
