+++ packages/sign-in-with-stacks/src/plugins/better-auth/types.ts
