+++ packages/sign-in-with-stacks/src/plugins/index.ts
@@ -1,0 +1,1 @@
+export * from "./better-auth/index.js";
