+++ packages/sign-in-with-stacks/tsdown.config.ts
@@ -1,7 +1,12 @@
 import { defineConfig, type UserConfig, type UserConfigFn } from "tsdown";
 
 const config: UserConfigFn | UserConfig = defineConfig({
-  entry: ["src/index.ts"],
+  entry: [
+    "src/index.ts",
+    "src/plugins/index.ts",
+    "src/plugins/better-auth/index.ts",
+    "src/plugins/better-auth/client.ts",
+  ],
   sourcemap: true,
   clean: true,
   dts: true,
