--- packages/siws-better-auth/package.json
@@ -1,60 +1,0 @@
-{
-  "name": "siws-better-auth",
-  "version": "0.0.1",
-  "license": "MIT",
-  "description": "better auth plugin for Sign-in with Stacks",
-  "keywords": [
-    "stacks",
-    "siws",
-    "sign-in-with-stacks",
-    "better-auth"
-  ],
-  "repository": "pradel/sign-in-with-stacks",
-  "funding": "https://github.com/sponsors/pradel",
-  "type": "module",
-  "sideEffects": false,
-  "publishConfig": {
-    "access": "public",
-    "provenance": true
-  },
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
-    },
-    "./client": {
-      "types": "./dist/client.d.ts",
-      "default": "./dist/client.js"
-    },
-    "./package.json": "./package.json"
-  },
-  "files": [
-    "dist",
-    "src/**/*.ts",
-    "!src/**/*.test.ts"
-  ],
-  "scripts": {
-    "build": "tsdown",
-    "check-types": "tsc --noEmit",
-    "dev": "tsdown --watch",
-    "lint": "oxlint .",
-    "test": "vitest"
-  },
-  "devDependencies": {
-    "@tsconfig/recommended": "1.0.13",
-    "@types/node": "24.10.1",
-    "better-auth": "1.4.6",
-    "publint": "0.3.16",
-    "sign-in-with-stacks": "workspace:*",
-    "tsdown": "0.17.3",
-    "typescript": "5.9.3",
-    "vitest": "4.0.15"
-  },
-  "peerDependencies": {
-    "better-auth": "^1.3.0",
-    "sign-in-with-stacks": "^0.1.0",
-    "zod": "^3.0.0 || ^4.0.0"
-  }
-}
