--- packages/siws-better-auth/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "extends": "@tsconfig/recommended/tsconfig.json",
-  "compilerOptions": {
-    "target": "ESNext",
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext"
-  },
-  "include": ["**/*.ts"],
-  "exclude": ["dist", "node_modules"]
-}
